kafka/service1/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
kafka-python==2.0.2
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.20.0
six==1.16.0
typing_extensions==4.1.1
